fix win warnings

+++ NAM/lstm.cpp
@@ -104,12 +104,12 @@ void lstm::LSTM::process(NAM_SAMPLE* input, NAM_SAMPLE* output, const int num_fr
     if (this->_stale_params)
     {
         for (std::unordered_map<std::string, double>::iterator it = this->_params.begin(); it != this->_params.end(); ++it)
-            this->_input_and_params[this->_parametric_map[it->first]] = it->second;
+            this->_input_and_params[this->_parametric_map[it->first]] = (float)it->second;
         this->_stale_params = false;
     }
     // Process samples, placing results in the required output location
     for (size_t i = 0; i < num_frames; i++)
-        output[i] = this->_process_sample(input[i]);
+        output[i] = this->_process_sample((float)input[i]);
 }
 
 float lstm::LSTM::_process_sample(const float x)
